pyproject.toml
[project]
name = "gemini"
version = "0.1.0"
description = "building an AI agent in Python"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
